test-requirements.txt
py
pytest
pytest-cov
pytest-django
